Tune PID for bed and extruder, calibrate extruder
Improves temperature control for the heated bed and extruder. Heating
happens more quickly and is more stable.

Fixes an issue where the extruder was overextruding, causing the hot end
to scrape and dislodge prints within the first few layers. It remains to
be seen if this is still an issue for taller prints.