Add auto-bed-leveling and adjust temp settings
Adds auto-bed-leveling with servo arm and z-probe on the z-max endstop
pin. Needs to be fine tuned, most likely with z-offset from the extruder
because it will scrape the bed in the current configuration.

Re-adjusts PID loop for heated bed and changes thermal runaway settings.
Thermistor on heated bed was reaseated resulting in this. Thermal
runaway still falsely triggers so settings may need to be adjusted
further (in particular it seems to be when filament is extruded over the thermistor). Otherwise temp has to be lowered and raised manually on the first layer.